% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{newhouse_quarter_report}
\alias{newhouse_quarter_report}
\title{report of stats about newly building house price}
\usage{
newhouse_quarter_report(
  reti_data,
  title = NULL,
  author = "",
  output_dir = "./"
)
}
\arguments{
\item{reti_data}{data.frame of reti}

\item{title}{title string}

\item{output_dir}{output directory path}
}
\description{
新築分譲住宅に関する取引価格の推移レポートを作成します.
}
