% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{render_wrapper}
\alias{render_wrapper}
\title{rmarkdown render wrapper}
\usage{
render_wrapper(
  df,
  template_file,
  title = "",
  author = "",
  output_file = "",
  output_dir = "handouts/"
)
}
\arguments{
\item{df}{data.frame 主にretiデータ}

\item{template_file}{rmarkdown file as template}

\item{title}{title string}

\item{author}{auther string}

\item{output_file}{output file name}

\item{output_dir}{output directory path デフォルトはカレントのhandouts}
}
\description{
自作のテンプレートファイルを使って、レポートを作ります。
出力ファイル名を省略すると、市名列から自動で補われます。
}
