% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_total_and_alup.R
\name{repoco_allresidence_quarter}
\alias{repoco_allresidence_quarter}
\title{report total and assumed land unit price about allresidence}
\usage{
repoco_allresidence_quarter(
  reti_data,
  title = "",
  author = "",
  output_file = "",
  output_dir = "./",
  cost = 190000,
  durability = 45
)
}
\arguments{
\item{reti_data}{data.frame of reti}

\item{title}{title string}

\item{author}{auther string}

\item{output_file}{output file name}

\item{output_dir}{output directory path}

\item{cost}{Int rebuilding unit cost}

\item{durability}{Int durabirity year}
}
\description{
report change of total and assumed land unit price
about all residence.
}
