% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_jirei.R
\name{repoco_jirei_tinyreport}
\alias{repoco_jirei_tinyreport}
\title{make tiny jirei report pdf}
\usage{
repoco_jirei_tinyreport(reti_jire_data, output_file = "", output_dir = "./")
}
\arguments{
\item{reti_jire_data}{data.frame with jirei data}

\item{output_file}{output file path}

\item{output_dir}{putput dir path}
}
\description{
This function is experimental.
}
