% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{repoco_newhouse_report}
\alias{repoco_newhouse_report}
\title{report of stats about newly building house price}
\usage{
repoco_newhouse_report(
  reti_data,
  title = "",
  author = "",
  output_file = "",
  output_dir = "./"
)
}
\arguments{
\item{reti_data}{data.frame of reti}

\item{title}{title string}

\item{author}{auther string}

\item{output_file}{output file name}

\item{output_dir}{output directory path}
}
\description{
新築分譲住宅に関する取引価格の推移レポートを作成します.
}
