% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_total_and_alup.R
\name{repoco_office_quarter}
\alias{repoco_office_quarter}
\title{report of total and assumed land unit price about office}
\usage{
repoco_office_quarter(
  reti_data,
  title = "",
  author = "",
  output_file = "",
  output_dir = "./",
  cost = 2e+05,
  durability = 50
)
}
\arguments{
\item{reti_data}{data.frame of reti}

\item{title}{title string}

\item{author}{auther string}

\item{output_file}{output file name}

\item{output_dir}{output directory path}

\item{cost}{Int rebuilding unit cost}

\item{durability}{Int durabirity year}
}
\description{
report change of total and assumed land unit price
about office building
}
