% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_line_graph.R
\name{rolling_mean_linegraph}
\alias{rolling_mean_linegraph}
\title{make line graph of simple mean change}
\usage{
rolling_mean_linegraph(df, title = "", keta = 3, unit = "千円")
}
\arguments{
\item{df}{data.frame of reti}

\item{title}{表のタイトル}
}
\description{
移動平均の推移に関する折れ線グラフを作る
}
