% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table.R
\name{simple_mean_table}
\alias{simple_mean_table}
\title{make table of simple mean change}
\usage{
simple_mean_table(df, title = "", keta = 3, unit = "千円")
}
\arguments{
\item{df}{data.frame of reti}

\item{title}{表のタイトル}

\item{keta}{平均額の省略桁}

\item{unit}{平均額の単位}
}
\description{
平均の推移に関する簡易な表を作る
}
