% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_report.R
\name{render_wrapper}
\alias{render_wrapper}
\title{rmarkdown render wrapper}
\usage{
render_wrapper(
  df,
  title,
  template_file,
  output_file = NULL,
  output_dir = "handouts/"
)
}
\arguments{
\item{df}{data.frame}

\item{title}{title string}

\item{template_file}{rmarkdown file as template}

\item{output_file}{output file name}

\item{output_dir}{output directory path}
}
\description{
call rmarkdown render function with template.
}
